We are done! Congratsgit add -A!

+++ capstone/tenmo-server/src/main/java/com/techelevator/tenmo/dao/UserDao.java
@@ -13,7 +13,7 @@ public interface UserDao {
 
     User findByUsername(String username);
 
-    int findIdByUsername(String username);
+    long findIdByUsername(String username);
 
     boolean create(String username, String password);
 }
